moved serialize functionality to Serialize.js plugin; define basic schema
